+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [1.0.17 - 1.0.17+1]
+
+- Format bubble value and `updatePositionOnTap` were added. Thanks @ltdangkhoa.
+
+## [1.0.16]
+
+- `dataPointFillColor` and `dataPointStrokeColor` were added in BezierLine.
+- `verticalLineFullHeight` was added in `BezierChartConfig`.
+- `onScaleChanged` was added in `BezierChart` to notify when the scale was changed.
+- Issue related to updating the chart with new data fixed. https://github.com/aeyrium/bezier-chart/issues/62
+- Web support was added
+
+## [1.0.15]
+
+- Fixed a bug related to `BezierChartScale.WEEKLY`
+- new property added `bezierChartAggregation` for `BezierChartConfig`
+- if datapoints now overlap the chosen aggregation method is used -> Supported are: `AVERAGE`, `SUM`, `FIRST`, `COUNT`
+
+## [1.0.14]
+
+- Fixed a bug related to `stepsYAxis` value
+- new property added `displayDataPointWhenNoValue` for `BezierChartConfig`
+- new property added `physics` for `BezierChartConfig`
+- Added `bubbleLabelValueBuilder` and `bubbleLabelDateTimeBuilder` callbacks
+
 ## [1.0.13]
 
 - `selectedValue` was added for `BezierChartScale.CUSTOM`